backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
opencv-python==4.10.0.84
numpy==2.0.0
Pillow==11.0.0